% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.ctreeMI}
\alias{print.ctreeMI}
\title{Print Method for ctreeMI Objects}
\usage{
\method{print}{ctreeMI}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ctreeMI"}, as returned by
\code{\link[=ctree_stacked]{ctree_stacked()}}.}

\item{...}{Further arguments passed to the partykit print method.}
}
\value{
\code{x}, invisibly.
}
\description{
Prints a summary of the stacked-imputation settings used to fit the
tree, followed by the standard \code{\link[partykit:ctree]{partykit::ctree()}} output.
}
