% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_table.R
\name{print.ctreeMI_report}
\alias{print.ctreeMI_report}
\title{Print a Methods Paragraph}
\usage{
\method{print}{ctreeMI_report}(x, width = 76, ...)
}
\arguments{
\item{x}{A \code{"ctreeMI_report"} object from \code{\link[=report_ctreeMI]{report_ctreeMI()}}.}

\item{width}{Wrapping width in characters.}

\item{...}{Currently unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Methods Paragraph
}
