% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_table.R
\name{print.ctreeMI_nodes}
\alias{print.ctreeMI_nodes}
\title{Print a Node Table}
\usage{
\method{print}{ctreeMI_nodes}(x, ...)
}
\arguments{
\item{x}{A \code{"ctreeMI_nodes"} object from \code{\link[=node_table]{node_table()}}.}

\item{...}{Passed to \code{\link[=print.data.frame]{print.data.frame()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Node Table
}
